+++ packages/ui-components/web/src/index.ts
@@ -5,7 +5,7 @@
 // // Components
 export { default as Button } from './Components/Button';
 // export { default as FileInput } from './Components/FileInput';
-// export { default as Image } from './Components/Image';
+export { default as Image } from './Components/Image';
 export { default as InputWithLabel } from './Components/InputWithLabel';
 export { default as LabeledToggle } from './Components/LabeledToggle';
 export { default as Panel } from './Components/Panel';